+++ CHANGELOG.md
@@ -6,9 +6,19 @@
 
 ### Добавлено
 
--
+- Добавлены endpoints для telebot [@fringer2423](https://github.com/fringer2423)
+- Добавлен Serializer для Telebot
+- Добавлен endpoint для чтения информации об одном филиале
+- Добавлен Serializer для Connect
+- Добавлены endpoints для Connect
 
 ### Изменено
+
+- Обновлена документация по новым endpoints [@fringer2423](https://github.com/fringer2423)
+- Обновлена документация по слою бизнес логики
+- Refactor endpoints
+- Доработана документация по API
+- Refactor code
 
 ### Удалено